Update Icon and added Audio
I update all of the project icons with the gunlinger skull. I also have
added in bullet sounds the play when the dueling can begin and when the
duel ends. I also added in a highNoon background music for the app.
NOTE: this music is purely for rec purposes and cannot be used if this
game is to be published, unless permision is obtained.